conf -> am.conf statistics -> am.stat idleMessage nicer (in my opinion) IO::timeXXX,IO::heapAllocated -> am.stat long -> unsigned int new const: MB/KB new fun:   void printMem(FILE *,char *,double) // print memory in Bytes/KByte or MB   void printTime(FILE *,char *,unsigned int) // print time in ms or sec statusMessage obsolete gc: getRealSize inlined by hand :-)

+++ platform/emulator/trail.hh
@@ -73,7 +73,7 @@ public:
 
   void popMark() {
     Assert(lastMark == tos-1);
-    lastMark -= (unsigned long) Stack::pop();
+    lastMark -= (unsigned int) Stack::pop();
   }
 
   int chunkSize()     { return (tos-1-lastMark)/2; }
